Create a BoundaryConfig Class
This class will hold the data of a single entry in
'./boundaries/boundary-config.json'.
At the moment it has a single method: 'as_dict' which returns an
OrderedDict, which can be serialized as json.
NB the 'boundary' keyword argument is the name that the directory and
layer will have in the commons directory.