Begin Create Config script implementation
Add a script that implements create-config  command
This implements step 1:
1. Creates the directory:`./boundaries/build/<name-of-boundary\>`
Still to go is:
2. Create an entry in boundaries/boundary-config.json filling in:
    - country
    - boundary
    - source-metadata
        - name
        - crs
        - encoding

To test this there is no 'tests/functional_tests.py'
The ambition is that this functional test file will replicate the
whole process of adding boundaries to a country.
This should serve the dual purpose of testing the package as well as
helping to identify ways of improving the automation process.

Add a functional test to check two create-config won't overwrite
an existing boundary of the same name.

Add create-config to setup.py